A1: Fix a WB bug
The WB stage would be fed data from the EX2 stage, whereas it should get
its data from the EX3 stage. Furthermore, the data is now fed
asynchronously from the EX3 stage to the write ports of the register
files (which are clocked). This is safe since the EX3/WB stages can
never be stalled (this has to be revisited once we implement interrupts
though).